Perform fill if db is empty
