+++ packages/graph-node/src/watcher.ts
@@ -569,3 +569,22 @@ export const getGraphDbAndWatcher = async (
     graphWatcher
   };
 };
+
+export const getStartBlock = (dataSources: any[]): number => {
+  const filteredContracts = dataSources.filter(contract => {
+    return (
+      contract.kind === 'ethereum/contract' &&
+      contract.source.startBlock !== undefined
+    );
+  });
+
+  if (filteredContracts.length === 0) {
+    return 0;
+  }
+
+  const minStartBlock = Math.min(
+    ...filteredContracts.map(contract => contract.source.startBlock)
+  );
+
+  return minStartBlock;
+};
